move the url_character function to WebCharacter.py
